Fixing parser and tests
The data loader can now load the CSV file - since the input is not
properly formatted, fields can contain commas, the data may be corrupt.

Added auto id generation for Server DAO - this will be needed to be
improved when changing to a different database.

Added PropertySource for the ServerService, so that the tests will run.

Changed findByHostName test so that it will pass regardless of the
order of the tests.